requirements.txt
pandas
sqlalchemy
scikit-learn
xgboost
joblib
sentence-transformers
faiss-cpu
transformers
langchain
fastapi
uvicorn
streamlit
pytest
flake8
requests